one problem solved!
There was a problem with the connection between the time points and objects. I had made a mistake. When for example, the time window was between 1000 and 2500, instead of corresponding objects, I used to take objects between 0 and 1500 (2500-1000)!!!
to do(s):
work on time windows of seeing each object instead of the number of frames, because the timing of frames is not constant!